requirements.txt
aiogram>=3.0.0
tortoise-orm>=0.19.0
aerich>=0.7.0
python-dotenv>=0.19.0
asyncpg>=0.27.0
aiosqlite==0.17.0
